Rehydrated DRYness
